# lib/api-zod/src/generated/types/submitResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitResponse {
  success: boolean;
  spreadsheetId?: string;
  message: string;
}